requirements.txt
streamlit
langchain-groq
langchain-community
langchain-core
faiss-cpu
sentence-transformers
python-dotenv
huggingface-hub